Get rid of remaining circular dependencies
