+++ extension/package.json
@@ -11,8 +11,8 @@
     "coverage": "npm run build:hooks && vitest --coverage",
     "test:playwright": "vitest --config vite.config.playwright.ts",
     "test:default": "vitest --config vite.config.ts",
-    "lint": "npx eslint . --fix && npx prettier --write .",
-    "lint:check": "npx eslint . && npx prettier --check ."
+    "lint": "dpdm src/background.ts --no-warning --no-tree --exit-code circular:1 && npx eslint . --fix && npx prettier --write .",
+    "lint:check": "dpdm src/background.ts --no-warning --no-tree --exit-code circular:1 && npx eslint . && npx prettier --check ."
   },
   "author": "Giulio B",
   "license": "MIT",
@@ -23,6 +23,7 @@
     "@types/firefox-webext-browser": "^143.0.0",
     "@vitest/browser": "^4.1.8",
     "@vitest/coverage-v8": "^4.1.8",
+    "dpdm": "^4.2.0",
     "eslint": "^10.4.1",
     "eslint-plugin-simple-import-sort": "^13.0.0",
     "globals": "^17.6.0",
